fix(ci): drop `--` separator + split video-on describe so e2e workflow finds tests (#281)
The previous run failed with "No tests found." for two compounding reasons,
both of which are addressed here.

1. **`--` separator in the `pnpm test:e2e` invocation.** Playwright's CLI
   treats `--` as a positional test-filter argument, not a Unix-style
   end-of-options marker. With pnpm 10 forwarding trailing args verbatim,
   `pnpm test:e2e -- --project=chromium` expanded to
   `playwright test -- --project=chromium`, which matches zero specs and
   exits 1. Dropping the `--` lets `--project=chromium` reach Playwright as
   an option.

2. **`test.use({ video: "on" })` inside a `describe` group.** Switching
   the video sink mid-suite forces a new worker, which Playwright disallows
   inside `test.describe`. Playwright surfaced this as
   "Cannot use({ video }) in a describe group" during test-collection,
   which silently zeroes out the file's test count and tips the whole
   suite into "0 tests in 0 files." Extracting the popover describe to its
   own file (`e2e/month-calendar-overflow-popover.spec.ts`) with
   file-level `test.use({ video: "on" })` preserves the per-describe
   video intent without touching the rest of `calendar.spec.ts`.

After both fixes, `pnpm exec playwright test --project=chromium --list`
enumerates 283 tests across 27 files locally (was 0).

Verification: `pnpm lint:fix && pnpm format:fix && pnpm check-types && pnpm test`
all green (2019/2019 vitest passing).
